# internal/services/api/docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  MessageError:
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: website-monitor/internal/services/api/docs
  WebSite:
    properties:
      Id:
        format: int64
        type: integer
      LastChecked:
        format: date-time
        type: string
      LastUpdated:
        format: date-time
        type: string
      Name:
        type: string
      Status:
        $ref: '#/definitions/WebsiteStatus'
      Url:
        type: string
    type: object
    x-go-package: website-monitor/internal/domain/entities
  WebsiteAddCommand:
    properties:
      name:
        description: website name
        maxLength: 255
        type: string
        x-go-name: Name
      url:
        description: website valid url
        maxLength: 555
        type: string
        x-go-name: Url
    required:
    - name
    - url
    type: object
    x-go-package: website-monitor/internal/domain/application/use_cases
  WebsiteAddResponse:
    properties:
      Id:
        format: int64
        type: integer
      Name:
        type: string
      Status:
        $ref: '#/definitions/WebsiteStatus'
      Url:
        type: string
    type: object
    x-go-package: website-monitor/internal/domain/application/use_cases
  WebsiteStatus:
    format: int64
    type: integer
    x-go-package: website-monitor/internal/domain/types
info:
  description: the purpose of this api is to provide a monitoring service for websites
  termsOfService: there are no TOS at this moment, use at your own risk we take no
    responsibility
  title: Website Monitor API
  version: 0.0.1
paths:
  /websites:
    get:
      description: Return a list of website from the application
      operationId: ListWebsites
      responses:
        "200":
          $ref: '#/responses/listWebsiteResponse'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - websites
    post:
      description: Add new website in application
      operationId: AddWebsite
      parameters:
      - description: Payload to add new website in application
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/WebsiteAddCommand'
      responses:
        "201":
          $ref: '#/responses/websiteAddResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
        "501":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - websites
  /websites/1:
    get:
      description: Return a website from the application
      operationId: GetWebiste
      responses:
        "200":
          $ref: '#/responses/websiteResponse'
        "404":
          $ref: '#/responses/notFoundResponse'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - websites
produces:
- application/json
responses:
  badRequestResponse:
    description: BadRequest message error returned as string
    schema:
      $ref: '#/definitions/MessageError'
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/MessageError'
  listWebsiteResponse:
    description: Data structure representing a list of website
    schema:
      items:
        $ref: '#/definitions/WebSite'
      type: array
  notFoundResponse:
    description: Not found message error returned as string
    schema:
      $ref: '#/definitions/MessageError'
  websiteAddResponse:
    description: Data structure representing website added
    schema:
      $ref: '#/definitions/WebsiteAddResponse'
  websiteResponse:
    description: Data structure representing a  websites
    schema:
      $ref: '#/definitions/WebSite'
schemes:
- http
swagger: "2.0"